+++ src/main/java/command/ByeCommand.java
@@ -1,6 +1,10 @@
-package main.java;
+package main.java.command;
 
-public class ByeCommand implements Command{
+import main.java.storage.Storage;
+import main.java.storage.TaskList;
+import main.java.ui.Ui;
+
+public class ByeCommand implements Command {
     @Override
     public void execute(TaskList taskList, Storage storage, Ui ui) {
         ui.printGoodbye();
