+++ src/main/java/command/Command.java
@@ -1,4 +1,8 @@
-package main.java;
+package main.java.command;
+
+import main.java.storage.Storage;
+import main.java.storage.TaskList;
+import main.java.ui.Ui;
 
 public interface Command {
     public void execute (TaskList taskList, Storage storage, Ui ui);
