+++ src/main/java/command/CommandEnum.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.command;
 
 public enum CommandEnum {
     BYE,
