+++ src/main/java/command/DeadlineCommand.java
@@ -1,4 +1,9 @@
-package main.java;
+package main.java.command;
+
+import main.java.task.Deadline;
+import main.java.storage.Storage;
+import main.java.storage.TaskList;
+import main.java.ui.Ui;
 
 import java.time.LocalDate;
