+++ src/main/java/command/ListCommand.java
@@ -1,4 +1,8 @@
-package main.java;
+package main.java.command;
+
+import main.java.storage.Storage;
+import main.java.storage.TaskList;
+import main.java.ui.Ui;
 
 public class ListCommand implements Command {
     @Override
