+++ src/main/java/command/UnmarkCommand.java
@@ -1,4 +1,9 @@
-package main.java;
+package main.java.command;
+
+import main.java.storage.Storage;
+import main.java.task.Task;
+import main.java.storage.TaskList;
+import main.java.ui.Ui;
 
 public class UnmarkCommand implements Command {
     private final int taskNumber;
