+++ src/main/java/exception/WindException.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.exception;
 
 public class WindException extends RuntimeException {
   public WindException(String message) {
