+++ src/main/java/task/Event.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.task;
 
 public class Event implements Task {
     private final String description;
