+++ src/main/java/task/Task.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.task;
 
 public interface Task {
